В index.html у тега script заменил одинарные кавычки на двойные Объект Numbers перенёс из util.js в data.js Фн. mockDataGenerate добавил параметр count В JSDoc добавил недостающие пробелы
