+++ js/main.js
@@ -1,4 +1,5 @@
 import { mockDataGenerate } from '/js/mock/data.js';
 
+const SIMILAR_PHOTO_DATA_COUNT = 5;
 
-window.console.log(mockDataGenerate());
+mockDataGenerate(SIMILAR_PHOTO_DATA_COUNT);
